+++ frontend/src/components/FormToolkit/index.tsx
@@ -1,11 +1,14 @@
 import SearchInput from "./SearchInput";
 import Input from "./Input";
+import Button from "./Button";
 
 const FormToolkit = {
   /** 2023/07/01 - 검색 input 컴포넌트 - by 1-blue */
   SearchInput,
   /** 2023/07/05 - "react-hook-form" 전용 input 컴포넌트 - by 1-blue */
   Input,
+  /** 2023/07/05 - button 컴포넌트 - by 1-blue */
+  Button,
 };
 
 export default FormToolkit;
